move out df building
